Switch from tinydb/joblib to json files
tinydb was having a lot of issues with not writing out results. I
thought I had resolved them all, but it turns out it would randomly
clear the .json database!  Probably has to do with some non-thread-safe
issue; I tried with and without joblib and including the tinyrecord
extension which is supposed to make it thread-safe, but to no avail.
So, we'll go with a straightforward solution: Just write out .json
files in a flat results directory.